Add QEMU-based test workflow
Only debian-bullseye-arch64 is supported initially. Packages are
fetched and cached outside of docker/QEMU to save time and to
avoid out of memory errors when cargo is fetching packages.

Note that care must be taken to ensure that the UID of the user
used by Github Actions match the UID of the docker user to avoid
permission errors when accessing files in a mounted volume..