`fn gather_{left,top}_partition_prob`: Make `r#in` args and `pc` var (from call) array refs.
`pc` was conditionally initialized to a valid ptr previously, or otherwise null.
The condition for checking if it's null was somewhat unrelated, so it'd be hard to make in an `Option`.
Since `pc` is `&mut [u16; 16]`, it was easier just to initially set it to
a zeroed array ref (`&mut Default::default()`) instead, which would also eliminate any extra `Option` checks.