`fn dav1d_decode_frame_init_cdf`: Make `f.ts`, `f.frame_thread.tile_start_off` slices before iterating over them.
This is very tricky with `f.frame_thread.tile_start_off` since it's only conditionally valid.
Using `iter::repeat(0)` as the alternative iterator doesn't work since they're different types,
so I make the slice len `0` in that case and chain it with `iter::repeat(0)`.

I'm not sure if this is bad for perf.