`struct Dav1dContext`: Deduplicate (last one and all its circular dependents).
`struct Dav1dContext` circularly contains references to:
* `struct Dav1dFrameContext`
  * `struct Dav1dFrameContext_bd_fn`
    * `type recon_b_intra_fn`
    * `type recon_b_inter_fn`
    * `type filter_sbrow_fn`
    * `type backup_ipred_edge_fn`
    * `type read_coef_blocks_fn`
* `struct Dav1dTaskContext`
which are also all deduplicated in the same commit since they are self-referential to each other.

Since `struct Dav1dContext` is a publically incomplete type in C,
we (for now) make it `pub` with a `pub use` unlike the rest of `src::internal`.
The correct way to expose such an opaque type is through type alias impl trait (TAIT),
but that's still unstable, so this should suffice now until we've progressed further.