`enum Rav1dError`: Use the cross-platform `libc::E*` `const`s for the discriminant values.
This also requires making `Rav1dError` `#[repr(u8)]`.

Furthermore, previously `92` was hardcoded for `ENOPROTOOPT` during transpilation,
but its value is platform dependent, so using `libc::ENOPROTOOPT` over hardcoding `92`
fixes the error values for `macOS`/`iOS`/`*bsd` OSes, where it's `42`,
and for Windows, where it's 123.
Similarly, `EAGAIN` is `11` on Linux and `35` on macOS.

Thus, we also now symbolicate the errno values in `tools/*` and `tests/*`
to use the `libc` `const`s so that they match the updated `Rav1dError` discriminants.