`struct Av1RestorationUnit`: Make `Copy` and change fields and vars to store by value, not pointer.
`Av1RestorationUnit` is only 10 bytes, just 2 more than the 8 bytes of ptr, which also comes with indirection.
Thus, this is likely more performant, and it get rid of nasty lifetimes and shared mutability.

The `lr` array var of `*const Av1RestorationUnit` `const` and just used locally, so the change there is simple.

The `lr_ref` array field is `*mut Av1RestorationUnit`, so `mut` and a field.
However, the `mut` is not needed, so the change becomes simple.