`build.rs`: Don't compile `avx` asm for `x86`.
`avx` (`avx2`, `avx512`) are only available on `x86_64`, not `x86`,
so we don't need to compile them on `x86`,
as that'll only take longer to compile and bloat the binaries.

`dav1d` doesn't seem to do this for some reason,
but in C, it gates `avx` asm calls behind `x86_64`, as we do, too.

For `ipred16`, `ipred16_avx2` defines a couple constants used by `ipred16_sse`,
so we have to compile `ipred16_avx2` on `x86`, still.
This is presumably a bug, and I'll try to upstream a fix for this.