`fn {generate_grain_uv,fguv_32x32xn}_{rust,c_erased}`: Put `#[inline(never)]` on the `fn *_c_erased`s, not the `fn *_rust`s.
The C `NOINLINE`s were originally on the `*_c` functions, which are called through `fn` ptrs and makes sense.
The `fn *_rust`s are called only by the `fn *_c_erased`s, which are called through `fn` ptrs,
so we were accidentally forcing an extra `fn` call due to mixing this up.