`mod {cdef,{cdef,lf,lr}_apply}_tmpl_{8,16}`: Always compile both and dead-code eliminate.
Since these bitdepth-specific `fn`s are Rust `fn`s, not asm `fn`s, we can't use `bd_fn!` in the same way,
so when `mod recon` got deduplicated, it broke compiling for a single bitdepth.
If we just compile both bitdepths for these `mod`s and rely on dead code elimination, it works again.
And once we deduplicate these `mod`s, it will all be for naught since they'll be generic instead.