`fn dav1d_ipred_*_neon`: Type-erase `BitDepth` generics.
One `fn` has a `const pixel px` arg, so we can't type erase it
because it's not behind a ptr and so the two types are different sizes.
It should be manageable since it's called in a Rust neon `fn`, not through a `fn` ptr,
and so we can just do the dispatch manually without `bd_fn!`
(which requires the two `fn` types to be unified).