`struct Rav1dIntraPredDSPContext::intra_pred`: Replace `Option`s with a `initialized: bool` field (on `struct Rav1dDSPContext`) as they are always `Some` even if `.is_none()` is checked.
`.is_none()` appears to be checked so that it can lazily initialize DSPs,
relying on things being zero-initialized to begin with.
Ideally we should move initialization to DSP allocation so its not lazy,
but for now (and the way closest to the C),
we can use a single initialized flag instead of `Option`s on everything.