`Cargo.toml`, `.github/workflows/*`: Separate profiles into `dev`, `fast-dev`, `release`, `checked-release` and update CI accordingly.
Previously, `profile.dev` had `opt-level = 1` for fast development,
but that's bad for debugging (`profile.dev` is also confusingly `debug`).
This separates the two into `profile.dev`/`debug` for debugging
and `profile.fast-dev` for the fastest local development cycle.

This also splits up `profile.release` into `profile.release`
and `profile.checked-release`, which is `profile.release`
with runtime checks and minimal debug info.
This is run in CI now (as opposed to a mix of `profile.dev` and `profile.release`),
except for `profile.release`, which is used for benchmarking.

This also consolidates some of the GitHub Actions workflows
to more easily manage the profiles, features, as some things were missing.
Specifically,
* argon tests were not run with `debug_assertions`
* arm tests were not run with `debug_assertions` or `overflow_checks`
* toolchains weren't cached on arm