`struct Rav1dITUTT35`: Make `payload{,_size}` fields a `Box<[u8]>`.
This makes `Rav1dITUTT35` no longer a DST, which is very tricky,
at the cost of an extra layer of indirection,
but since the payload probably won't be accessed through the `Rav1dITUTT35` excessively,
I don't think this should be a problem.
Note that having a separate `Rav1dITUTT35` from `Dav1dITUTT35` allows us to do this easily.