`struct Rav1dThreadPicture::progress`: Make a `struct ThreadPictureProgress` of [`AtomicU32; 2]` that is always allocated inline since it's so small.
Previously, `progress` was either null or `*const [AtomicU32; 2]`,
allocated by the `extra` and `extra_ptr` in `fn picture_alloc_with_edges`
as part of the `pic_ctx` allocation.
But since `extra_ptr` is only ever the `*const [AtomicU32; 2]` or null,
we can just allocate it inline and avoid this tricky dynamically-sized allocation.

I made `ThreadPictureProgress` a newtype to `impl Clone` on it,
as this has to be done manually for atomics.
It also lets us name the `[0]` and `[1]` indices as `{block,pixel}_data` as the comments say.