`struct Rav1dThreadPicture::progress`: Make an `Option<Arc<_>>` separate from the `pic_ctx_context` `Rav1dRef`.
Previously, `progress` was part of the `pic_ctx_context` `Rav1dRef`,
but optionally using a DST.
To avoid that at the cost of a separate `Arc` allocation,
this moves `progress` to a separate `Option<Arc<_>>`.
The `pic_ctx_context` `Rav1dRef` is especially complex to `Arc`ify due to the `fn` ptrs, etc. in it,
so while this does incur an extra allocation cost
and an extra `Arc::clone` on `Rav1dThreadPicture::clone`, it should still be worth it.