`fn BitDepth::pxstride`: Implement generically over `stride` type, e.x. `isize` or `usize`.
This way we don't have to do `BD::pxstride(stride as usize) as isize`,
which is verbose and wrong for negative strides.
There may still be some places where strides are always positive, though, so it's kept generic.
This also lets us fix #636 gradually; this is the first major step in fixing #636.