`struct Rav1dFrameContext_task_thread`: Make Rav1dTask lists indexed
Rav1dTasks are stored in two arrays and linked
into a linked list by pointers. This change
refactors the lists into a `Rav1dTasks` structure
that links tasks by index instead of pointer.

We still access the Rav1dTasks structure with
unsafe interior mutability, because we would have
borrowing errors between the Rav1dFrameData
reference lifetime and the lifetime of borrows out
of the Rav1dTasks structure. This will be fixed
soon when we split `Rav1dFrameData` into an
immutable, interiorly mutable portion and a mutex
protected mutable portion.