`struct refmvs_frame::frm_hdr`: Replace raw ptr with `*const ()`.
This removes the `*const Rav1dFrameHeader` in the field,
passing it through the call stack.
But we can't change the field's layout since `*const refmvs_frame` is passed to asm,
so we make the field `*const ()` and `ptr::null()`.
This also fixes the problem of accidentally conflating
`*const Rav1dFrameHeader` and `*const Dav1dFrameHeader` in that field.