Make `DisjointMut::index()` safe
We were not adding any useful information at the numerous call sites for
`index()`, and the safety involved here is thread safety to avoid data
races, which we don't require for this application when handling pixel
data and other plain data collections. This change makes `index()` a
safe function and only requires safety comments for mutable access.