`struct FFISafe`: Add a type that lets us roundtrip a non-FFI safe type through an `extern "C" fn` call.
For non-FFI safe types, we get warnings if we pass it as is.
We can disable those warnings, but only on the whole `fn`,
and I'd like those checks to still be done on the other args.
This bypasses the check for a single type,
and it type-safely lets us cast it back in a Rust `fn`,
where FFI safety won't be a concern since caller and callee are both Rust.