`test.sh`: Set `RUST_LIB_BACKTRACE=0` if not overridden so `debug_assertions` tests don't take forever due to capturing a ton of backtraces.
We still want `RUST_BACKTRACE=1` to print a backtrace on panicking, as that's very useful and doesn't slow tests down.
But that also causes `Backtrace::capture()` to actually capture a backtrace,
and since we do that in every `DisjointMut::{index,index_mut}` call, it makes the tests extremely slow.
Setting `RUST_LIB_BACKTRACE=0` overrides this and makes the tests fast (enough) again.
And it can always be overridden again on the command line.