`mod msac`: Enable `sse2` and `neon` asm `fn`s when those features are statically available.
This aligns with what C currently does, and fixes our previous mistranslation of that behavior,
as the `sse2` and `neon` `fn`s were not used on `x86` and `armv7` even when enabled at compile time.
However, this does not align with the behavior elsewhere of dynamically detecting cpu features,
so the `sse2` `fn`s would not be used on `x86` unless specifically compiled with them.