`fn loopfilter_sb`: Make `mask` arg a `&[u32; 3]` array ref.
The uv `fn`s were called with `&[u32; 3]`s, but the y `fn`s were called with `&[u32; 4]`.
However, the 4th element was always zero, is never read in the fallback `fn`s,
and doesn't look like it's accessed by asm either (not even as a `u32x4` `|` or something).