`type Dav1dContext`: Change from `Rav1dContext` to `RawArc<Rav1dContext`.
This changes `Dav1dContext` from an opaque pointee type to an opaque pointer type.
This is different from the `DAV1D_API`, but since `Dav1dContext` is opaque,
we can do this from the Rust side, and it will match up with `Dav1dContext*` usages from C.
This allows us to make the `DAV1D_API`s simpler and safer, so it's useful,
and we can accurately model `Option<Dav1dContext>`.