`fn ipred_filter_rust`: Separate `p{5,6}` and `left` calculations for `topleft_in` and `dst` derived ptrs.
`left` and `p{5,6}` are calculated from both the `topleft_in` slice and `dst` picture data depending on the iteration.
This is trickier when `dst` becomes a `Rav1dPictureDataComponentOffset` due to the negative indexing.
So this changes to using an `if` to decide which version to calculate instead of storing them both in the same loop var.
This `if` will either be removed if the loop is unrolled a bit,
or mispredicted only once since it only uses `topleft_in` for the first iteration.
And also, this is in a fallback `fn`, so its perf is less important.