`fn CfSelect::{get,set}`: Remove and replace with direct indexing, slicing once in `fn decode_coefs`.
For the `CfSelect::Task` case, no slicing is done, the array is just used as is.
For the `CfSelect::Frame` case, the `DisjointMut` is sliced to get a `cf_len` slice.
Then `fn get` and `fn set` can directly index into this array with minimal bounds checks and no `match`.