`fn decode_coefs_class`: Replace simple branch w/ branchless logic.
Although the previous code would usually emit branchless `cmov`s,
this was not guaranteed and a branch was emitted in one the 16bpc `TxClass`es,
causing mispredictions and slowdowns.
This changes the logic to use only bit operations.

This effectively replaces a `test` and `cmov` with a `xor`, `and`, `xor`,
which should be similarly efficient.
In the branched case, it is far more performant.

Co-authored-by: Frank Bossen <fbossen@gmail.com>