`struct CaseSet`: Optimize by `match`ing on `len` one per `set_ctx`.
In C, `case_set` works by `switch`ing once on `len`, once per `set_ctx`,
but my `CaseSet` implementation in Rust accidentally switched to `match`ing on `buf.len()` each time.
This goes back to the original, optimal behavior.

To do this, `set_ctx` is made a rank-2 polymorphic "closure".
This does not exist in Rust, so it's emulated through a generic trait with a generic method.
This inner generic over `trait CaseSetter` is what allows
`fn CaseSet::one` to select the correct `CaseSetterN` at compile time.

However, this means that closures can't be used anymore, which is very annoying.
To partially remedy this, I added the `set_ctx!` macro,
which emulates the `set_ctx` closure as much as possible.
All captures (up vars in `rustc`) and their types must be declared.