Rearrange default cdf values to enable fewer copies
Create `CdfDefaultContext` structure that holds default cdf values
and closely mirrors the `CdfContext` structure.

When copying data from `CdfDefaultContext` to `CdfContext`, the
fields `coef`, `m`, `mi`, `mv`, and `kyfm` are copied. For `coef`,
one of the four set of values in `CdfDefaultContext` is copied.
For `mv`, `CdfDefaultContext` contains only one set of values for
`mv.comp` whereas `CdfContext` contains two. Hence, this section
of data is copied twice.

In the C code, memory copy operations are optimized to be done in
3 steps:
- `coef`
- `m`, `mi`, and first copy of `mv.comp`
- second copy of `mv.comp`, remainder of `mv`, and `kyfm`

This optimization is not done explicitely in Rust. However, a sufficiently
smart compiler should be able to produce code equivalent to the C one.