Elide a bounds check in get_lo_ctx
The function accesses level pairs (0, 1), (0, 2), (0, 3), (0, 4), (1, 0), (1, 1) and (2, 0)
for computing the result.

Notably, it does not access levels(1, 4). However, it does try to bounds check that pair
in an attempt to elide other bounds checks.

We can prove statically that access to the required pairs is safe. We cannot statically
prove that access to (1, 4) is safe. Doing the maths, I think this is because an access
to (1, 4) is not actually guaranteed to be safe.

Stop trying to help the compiler here.

This removes a jump to a bounds check failure handler from both versions of decode_coefs.