get_lo_ctx: min(4) is expensive, unroll array instead
This trades off memory usage (and processor cache efficiency) to get rid of
some really annoying assembly that LLVM generates.

Before:
```
        movzx edx, r13b
        cmp dl, 4
        mov edi, 4
        cmovae edx, edi
        movzx edx, dl
        movzx esi, bl
        cmp sil, 4
        cmovae esi, edi
        movzx esi, sil
                // /home/dja/dev/av1/rav1d/src/recon.rs : 507
                offset = ctx_offsets[cmp::min(y as usize, 4)][cmp::min(x as usize, 4)].get();
        lea rdx, [rdx + 4*rdx]
        add rdx, qword ptr [rsp + 168]
        movzx edx, byte ptr [rsi + rdx]
```

We have compare, conditonal move, compare, conditional move, calculate address,
load byte. The load byte is really hot in perf, not because the memory access is
all that slow, but because the memory access cannot begin until all the preceeding
instructions to calculate the address have finished, and they form a dependency
chain. You'd hope the compiler would spread them out so that it's not so bad for
latency, but it doesn't.

With this, we get an address calculation that doesn't use `cmov`s at all.

While a 5x5 little array is lovely and fits in a cache line, at least on my old x86-64
laptop it's much faster to do this: I see a further ~0.6% less time spent in decode_coefs
according to perf diff on the summer_nature_1080p.ivf file.