Replace inlined, repeatedly-bounds-checking zeroing with memset/bzero
It turns out both the H and V cases were already calling out to the standard library.

This is based on the arm64 assembler for macOS:

This is the original code for the H case. The compiler has noted that stride is 16, and so has
rearranged the code as follows:

```
  stride as usize * ((4 << slwh_zero) as usize + 2)
=              16 * ((4 << slwh_zero) as usize + 2)
=              16 * 4 << slwh_zero + 16 * 2
=              64 << slwh_zero + 32
```

Giving:

```
        mov w8, #64			# w8 = 64
                // /Users/danielaxtens/src/av1/rav1d/src/recon.rs : 866
                for i in 0..stride as usize * ((4 << slwh_zero) as usize + 2) {
        lsl w8, w8, w25 		# w8 = w8 << slwh_zero
        orr w1, w8, #0x20		# w1 = w8 + 32 (0x20)
                                        # (as (64 << n) > 32, we can use biwise-or to add)
                // /Users/danielaxtens/src/av1/rav1d/src/recon.rs : 867
                levels[i] = 0;
        mov x0, x24
        bl _bzero                       # bzero(levels, 64<<slwh_zero + 32)
```

Basically, the compiler's loop recognition has kicked in and it's detected a bzero
and rewritten it to be a call to bzero, which the C library can then optimise appropriately.
We know slwh_zero is no more than 3, and 64 << 3 + 32 = 512 + 32 = 544, which will always
be less than the size of the array, 1088, so there's no explicit bounds check.

V has exactly the same pattern with slightly different registers (slh vs slw).

TwoD, on the other hand, had a vectorised fast path, but the byte-by-byte part included a
bounds check on every loop iteration:

This is the first part:

```
        ldr x11, [sp, #152]
                // /Users/danielaxtens/src/av1/rav1d/src/recon.rs : 830
                stride = 4 << slh;
        lsl x15, x22, x11
        stur x15, [x29, #-136]
                // /Users/danielaxtens/src/av1/rav1d/src/recon.rs : 845
                shift = slh + 2;
        add w11, w11, #2
        ldr w15, [sp, #124]
                // /Users/danielaxtens/src/av1/rav1d/src/recon.rs : 866
                for i in 0..stride as usize * ((4 << slwh_zero) as usize + 2) {
        lsl w15, w22, w15
        orr w15, w15, #0x2
        lsl x15, x15, x11
        cmp x15, #33                    # if we have at least 33 bytes...
        b.hs LBB320_73                  # go to the fast path,
        mov x0, #0
        mov w16, #1
        b LBB320_75                     # otherwise go to the byte-by-byte loop
```

LBB320_73 is a vectorised zero loop, that writes in lots of 32 bytes.

Then there's a loop for doing byte-by-byte:

```
LBB320_75:
        add x17, x20, x21
        str x17, [sp, #48]
        umaddl x12, w12, w13, x14
        str x12, [sp, #144]
        ldur x12, [x29, #-136]
        sub w20, w12, #1
        add x12, x15, #1
LBB320_76:
        cmp x16, #1089     # <-------------- this is a bounds check!
        b.eq LBB320_205    # <-------------- this the branch to a panic!
        mov x13, x16
        strb wzr, [x24, x0]
                // /Users/danielaxtens/.rustup/toolchains/nightly-2025-05-01-aarch64-apple-darwin/lib/rustlib/src/rust/library/core/src/iter/range.rs : 764
                if self.start < self.end {
        add x16, x16, #1
        mov x0, x13
                // /Users/danielaxtens/src/av1/rav1d/src/recon.rs : 866
                for i in 0..stride as usize * ((4 << slwh_zero) as usize + 2) {
        cmp x12, x16
        b.ne LBB320_76
...

LBB320_205:
                // /Users/danielaxtens/src/av1/rav1d/src/recon.rs : 867
                levels[i] = 0;
Lloh3609:
        adrp x2, l_anon.af3a7bc02be78efe14397b07e69a5ff1.1558@PAGE
Lloh3610:
        add x2, x2, l_anon.af3a7bc02be78efe14397b07e69a5ff1.1558@PAGEOFF
        mov w1, #1088
        bl core::panicking::panic_bounds_check
```

Rust should be able to elide the bounds check here but clearly it is not doing so.
Rearrange the calculation (not strictly necessary but avoids some instructions)
and that leads to rust being able to avoid the bounds check. Then we can just use
.fill(0) everywhere: it goes to libc's memset implementation, which we trust to be
optimised for each platform.

This does appear to go faster: perf diff reports ~0.75% less time in decode_coefs
with this patch compared to the previous patch.