Use Dav1dSequenceHeader instead of Rav1dSequenceHeader
Remove Rav1dSequenceHeader which is basically a duplicate representation of the Dav1dSequenceHeader type. While the Rav1dSequenceHeader uses some more typical Rust types, theDav1dSequenceHeader belongs to the public API which needs to be kept so it the one that remains. Since both types represent the same concept it is not needed to have both of them.