perf: make `Mv` `align(4)`
This matches up with how it works in the C code, where it's a union with a u32.

It requires a new struct for inclusion in RefMvsTemporalBlock, but nothing we can't handle.

It improves the code in `add_spatial_candidate` a lot. On nightly on macOS, we see for example this:

```
ldrh  w11, [x3]
ldrh  w10, [x3, #2]
strh  w11, [sp, #8]
strh  w10, [sp, #10]
ldr   w8, [sp, #8]
```

becomes this:
```
ldr   w10, [x3]
```

And later on we see some branches replaced with conditional moves. We also see better code on x86.

`perf` on x86 shows a significant improvement in time spent in `add_spatial_candidate`.

I see an overall improvement around 1% on the Chimera files on a Windows system with an AMD Ryzen 5 2600 6-core processor.

I tested with `hyperfine`, with 6 threads (1 per core) in order to reduce system noise, but the figures are still very rough.