Remove the now-unnecessary `'next` block
When iterating tasks, we used to need to do certain things at the end of the loop.
We no longer need to do those, so we can get rid of a block and reduce indentation
and improve clarity.

Broken out from the other changes to make it clearer that the bulk of the change
is whitespace rather than substantive.