Take the task vector read lock less
Even though `.try_read().unwrap()` in `.index()` and `.len()` is guaranteed to succeed,
it's not free. Try to do it less.

This requries that we have slightly more involved handling for when we have to insert
a new task as a result of finding a task, but it means that we don't have to take
the read lock every iteration of the loop.