`struct Rav1dFrameContextTaskThreadPendingTasks`: remove unused code
This seems to have been removed a long time ago in 8f78cfe,
but we never realized that it was dead code since it was `pub`.
The newly released Rust 1.89 warns about this, however,
since it was never constructed either,
and our `cargo +stable clippy -- -D warnings` CI step uses the latest stable.